update is ok - WIP mes taches accomplies

+++ backend/app/controllers/tasksListController.js
@@ -37,6 +37,8 @@ const tasksListSchema = Joi.object({
     .falsy('false')
     .required(), 
     date_perso: Joi.date()
+    .min('now')
+    .required(),
 
 });
 
@@ -140,13 +142,13 @@ const tasksListController = {
         //* Update the tasksList
         try {
             console.log(">> tasksListContr l.144 req.body", req.body); // exemple -> { task_id: '26', move_id: 19, is_realised: false }
-            const tasksListValidation = tasksListSchema.validate(req.body); 
-            // if an error is found 
-            if (!!tasksListValidation.error) {
-                // abort and send error 400 : bad request
-                res.status(400).send(tasksListValidation.error); 
-            }
-            
+            // --------->>>> la validation rencontre un problème, la reprendre
+            // const tasksListValidation = tasksListSchema.validate(req.body); 
+            // // if an error is found 
+            // if (!!tasksListValidation.error) {
+            //     // abort and send error 400 : bad request
+            //     res.status(400).send(tasksListValidation.error); 
+            // }
             // form is valid !
             
             const selectedTaskInList = await TasksList.getByPk(req.params.moveId, req.params.taskId);
@@ -181,7 +183,7 @@ const tasksListController = {
             for (const prop in req.body) {
                 selectedTaskInList[prop] = req.body[prop]; 
             }
-            
+            console.log('l.183 tasklisController : req.body',req.body)
             // Execute request
             
             const updatedTask = await selectedTaskInList.updateTasksList(); 
@@ -253,7 +255,7 @@ const tasksListController = {
 
             // 2e -> pour supprimer la tache
             const storedTask = await Task.getTaskByPk(req.params.taskId);
-            console.log(">> l.195 storedTask",req.params.taskId);
+            console.log(">> l.195 storedTask.general_task",storedTask.general_task);
              // If no box was found 
              if (!storedTask ) {
                  // Abort and send error : 404 not found
@@ -267,19 +269,23 @@ const tasksListController = {
                      }
                  });
              }
-             // on supprime la tache
-             const successTask = await storedTask.delete();      
-             if (!successTask) {
-                 return res.status(500).send({
-                     statusCode : 500,
-                     message:  {
-                         en:"Something went wrong", 
-                         fr:"Quelque chose s'est mal passée"
-                     }
-                 });
+             // on supprime la tache si elle n'est pas générale
+             if (!storedTask.general_task) {
+                 const successTask = await storedTask.delete();      
+                 if (!successTask) {
+                     return res.status(500).send({
+                         statusCode : 500,
+                         message:  {
+                             en:"Something went wrong", 
+                             fr:"Quelque chose s'est mal passée"
+                         }
+                     });
+                 }
              }
 
             res.send(successTaskInList);
+            
+            
         } catch (err) {
             console.trace(err);
         }
